added timers and sample start

+++ alsuq/include/alsuq/stats/StatisticsTimer.hpp
@@ -3,12 +3,12 @@
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
@@ -57,9 +57,9 @@ private:
     std::string name;
     const std::shared_ptr<Statistics> statistics;
 
-    int statisticsTime = 0;
-    int combineTime = 0;
-    int finalizeTime = 0;
+    double statisticsTime = 0;
+    double combineTime = 0;
+    double finalizeTime = 0;
 };
 } // namespace stats
 } // namespace alsuq
